requirements.txt
# SDR Telegram Bot - filtros de água (SPIN Selling)
# Python 3.10+

# Telegram
python-telegram-bot>=21.0

# OpenAI (LLM, Whisper STT, TTS)
openai>=1.40.0

# Google Drive API
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# Ambiente e async
python-dotenv>=1.0.0

# Banco de dados
# SQLite: stdlib. Supabase REST: supabase. Postgres (connection string): psycopg2
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Message buffer (debounce)
redis>=5.0.0

# Platform backend (FastAPI) — usado no deploy Vercel da API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
# python-jose[cryptography] inclui cryptography, mas vamos explicitar para garantir a instalação no Vercel
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
bcrypt>=4.0.0
python-multipart>=0.0.6
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
httpx>=0.25.0
stripe>=8.0.0
